--- a/src/System.CommandLine.Benchmarks/CommandLine/Perf_Parser_Options_Bare.cs
+++ b/src/System.CommandLine.Benchmarks/CommandLine/Perf_Parser_Options_Bare.cs
@@ -19,7 +19,7 @@ namespace System.CommandLine.Benchmarks.CommandLine
         private string _testSymbolsAsString;
         private Parser _testParser;
 
-        private IEnumerable<Option> GenerateTestOptions(int count, IArgumentArity arity)
+        private IEnumerable<Option> GenerateTestOptions(int count, ArgumentArity arity)
             => Enumerable.Range(0, count)
                          .Select(i =>
                                      new Option($"-option{i}", arity: arity)