--- a/src/System.CommandLine.Tests/ParserTests.cs
+++ b/src/System.CommandLine.Tests/ParserTests.cs
@@ -1453,9 +1453,9 @@ namespace System.CommandLine.Tests
                    .Tokens
                    .Should()
                    .BeEquivalentTo(
-                       new Token("1", TokenType.Argument),
-                       new Token("2", TokenType.Argument),
-                       new Token("3", TokenType.Argument));
+                       new Token("1", TokenType.Argument, default),
+                       new Token("2", TokenType.Argument, default),
+                       new Token("3", TokenType.Argument, default));
         }
 
         [Fact]
@@ -1474,19 +1474,19 @@ namespace System.CommandLine.Tests
                    .Tokens
                    .Should()
                    .BeEquivalentTo(
-                       new Token("1", TokenType.Argument),
-                       new Token("2", TokenType.Argument),
-                       new Token("3", TokenType.Argument));
+                       new Token("1", TokenType.Argument, default),
+                       new Token("2", TokenType.Argument, default),
+                       new Token("3", TokenType.Argument, default));
             command.Parse("1 2 3 4 5")
                    .CommandResult
                    .Tokens
                    .Should()
                    .BeEquivalentTo(
-                       new Token("1", TokenType.Argument),
-                       new Token("2", TokenType.Argument),
-                       new Token("3", TokenType.Argument),
-                       new Token("4", TokenType.Argument),
-                       new Token("5", TokenType.Argument));
+                       new Token("1", TokenType.Argument, default),
+                       new Token("2", TokenType.Argument, default),
+                       new Token("3", TokenType.Argument, default),
+                       new Token("4", TokenType.Argument, default),
+                       new Token("5", TokenType.Argument, default));
         }
 
         [Fact]
@@ -1543,9 +1543,9 @@ namespace System.CommandLine.Tests
                    .Tokens
                    .Should()
                    .BeEquivalentTo(
-                       new Token("1", TokenType.Argument),
-                       new Token("2", TokenType.Argument),
-                       new Token("3", TokenType.Argument));
+                       new Token("1", TokenType.Argument, default),
+                       new Token("2", TokenType.Argument, default),
+                       new Token("3", TokenType.Argument, default));
         }
 
         [Fact]
@@ -1563,19 +1563,19 @@ namespace System.CommandLine.Tests
                    .Tokens
                    .Should()
                    .BeEquivalentTo(
-                       new Token("1", TokenType.Argument),
-                       new Token("2", TokenType.Argument),
-                       new Token("3", TokenType.Argument));
+                       new Token("1", TokenType.Argument, default),
+                       new Token("2", TokenType.Argument, default),
+                       new Token("3", TokenType.Argument, default));
             command.Parse("-x 1 -x 2 -x 3 -x 4 -x 5")
                    .FindResultFor(option)
                    .Tokens
                    .Should()
                    .BeEquivalentTo(
-                       new Token("1", TokenType.Argument),
-                       new Token("2", TokenType.Argument),
-                       new Token("3", TokenType.Argument),
-                       new Token("4", TokenType.Argument),
-                       new Token("5", TokenType.Argument));
+                       new Token("1", TokenType.Argument, default),
+                       new Token("2", TokenType.Argument, default),
+                       new Token("3", TokenType.Argument, default),
+                       new Token("4", TokenType.Argument, default),
+                       new Token("5", TokenType.Argument, default));
         }
 
         [Fact]