--- a/src/System.CommandLine/ArgumentArity.cs
+++ b/src/System.CommandLine/ArgumentArity.cs
@@ -9,9 +9,13 @@ using System.Diagnostics;
 
 namespace System.CommandLine
 {
-    /// <inheritdoc />
+    /// <summary>
+    /// Defines the arity of an option or argument.
+    /// </summary>
+    /// <remarks>The arity refers to the number of values that can be passed on the command line.
+    /// </remarks>
     [DebuggerDisplay("\\{{" + nameof(MinimumNumberOfValues) + "},{" + nameof(MaximumNumberOfValues) + "}\\}")]
-    public class ArgumentArity : IArgumentArity
+    public readonly struct ArgumentArity : IEquatable<ArgumentArity>
     {
         private const int MaximumArity = 100_000;
 
@@ -41,13 +45,30 @@ namespace System.CommandLine
 
             MinimumNumberOfValues = minimumNumberOfValues;
             MaximumNumberOfValues = maximumNumberOfValues;
+            IsNonDefault = true;
         }
 
-        /// <inheritdoc />
+        /// <summary>
+        /// Gets the minimum number of values required for an <see cref="IArgument">argument</see>.
+        /// </summary>
         public int MinimumNumberOfValues { get; }
 
-        /// <inheritdoc />
+        /// <summary>
+        /// Gets the maximum number of values allowed for an <see cref="IArgument">argument</see>.
+        /// </summary>
         public int MaximumNumberOfValues { get; }
+
+        internal bool IsNonDefault { get;  }
+
+        public bool Equals(ArgumentArity other) => 
+            other.MaximumNumberOfValues == MaximumNumberOfValues && 
+            other.MinimumNumberOfValues == MinimumNumberOfValues &&
+            other.IsNonDefault == IsNonDefault;
+
+        public override bool Equals(object obj) => obj is ArgumentArity arity && Equals(arity);
+
+        public override int GetHashCode()
+            => MaximumNumberOfValues ^ MinimumNumberOfValues ^ IsNonDefault.GetHashCode();
 
         internal static FailedArgumentConversionArityResult? Validate(
             SymbolResult symbolResult,
@@ -94,29 +115,29 @@ namespace System.CommandLine
         /// <summary>
         /// An arity that does not allow any values.
         /// </summary>
-        public static IArgumentArity Zero { get; } = new ArgumentArity(0, 0);
+        public static ArgumentArity Zero => new ArgumentArity(0, 0);
 
         /// <summary>
         /// An arity that may have one value, but no more than one.
         /// </summary>
-        public static IArgumentArity ZeroOrOne { get; } = new ArgumentArity(0, 1);
+        public static ArgumentArity ZeroOrOne => new ArgumentArity(0, 1);
 
         /// <summary>
         /// An arity that must have exactly one value.
         /// </summary>
-        public static IArgumentArity ExactlyOne { get; } = new ArgumentArity(1, 1);
+        public static ArgumentArity ExactlyOne => new ArgumentArity(1, 1);
 
         /// <summary>
         /// An arity that may have multiple values.
         /// </summary>
-        public static IArgumentArity ZeroOrMore { get; } = new ArgumentArity(0, MaximumArity);
+        public static ArgumentArity ZeroOrMore => new ArgumentArity(0, MaximumArity);
 
         /// <summary>
         /// An arity that must have at least one value.
         /// </summary>
-        public static IArgumentArity OneOrMore { get; } = new ArgumentArity(1, MaximumArity);
+        public static ArgumentArity OneOrMore => new ArgumentArity(1, MaximumArity);
 
-        internal static IArgumentArity Default(Type type, Argument argument, ISymbolSet parents)
+        internal static ArgumentArity Default(Type type, Argument argument, ISymbolSet parents)
         {
             if (type == typeof(bool))
             {