--- a/src/System.CommandLine/Help/HelpOption.cs
+++ b/src/System.CommandLine/Help/HelpOption.cs
@@ -36,6 +36,8 @@ namespace System.CommandLine.Help
 
         internal override Argument Argument => Argument.None();
 
+        internal override bool IsGreedy => false;
+
         public override bool Equals(object obj)
         {
             return obj is HelpOption;