… src/System.CommandLine.ApiCompatibility.Tests/ApiCompatibilityApprovalTests.System_CommandLine_api_is_not_changed.approved.txt
… src/System.CommandLine/Binding/FailedArgumentTypeConversionResult.cs
… src/System.CommandLine/Builder/CommandLineBuilderExtensions.cs
… src/System.CommandLine/Collections/AliasedSet.cs
… src/System.CommandLine/Collections/ISymbolSet.cs
… src/System.CommandLine/Collections/SymbolSet.cs
… src/System.CommandLine/Command.cs
… src/System.CommandLine/CommandLineConfiguration.cs
… src/System.CommandLine/Help/HelpBuilder.Default.cs
… src/System.CommandLine/Help/HelpBuilder.cs
… src/System.CommandLine/Help/VersionOption.cs
… src/System.CommandLine/IArgument.cs
… src/System.CommandLine/IArgumentArity.cs
… src/System.CommandLine/IMatchable.cs
… src/System.CommandLine/ISymbol.cs
… src/System.CommandLine/IdentifierSymbol.cs
… src/System.CommandLine/Invocation/SuggestDirectiveResult.cs
… src/System.CommandLine/Option.cs
… src/System.CommandLine/Option{T}.cs
… src/System.CommandLine/ParentNode.cs
… src/System.CommandLine/Parsing/CommandResult.cs
… src/System.CommandLine/Parsing/ImplicitToken.cs
… src/System.CommandLine/Parsing/OptionResult.cs
… src/System.CommandLine/Parsing/ParseOperation.cs
… src/System.CommandLine/Parsing/ParseResult.cs
… src/System.CommandLine/Parsing/ParseResultExtensions.cs
… src/System.CommandLine/Parsing/ParseResultVisitor.cs
… src/System.CommandLine/Parsing/StringExtensions.cs
… src/System.CommandLine/Parsing/SymbolResult.cs
… src/System.CommandLine/Parsing/SymbolResultExtensions.cs