@jonsequitur I am about to start my weekend now, so I am just opening the PR and providing benchmark results. On Monday I am going to describe why I've changed all of these things.

|        Method |      Args | Mean before | Mean after | Allocated before | Allocated  after|
|-------------- |---------- |------------:|-----------:|-----------------:|----------------:|
|  DefaultsSync | String[0] |   38.892 us |  30.004 us |         25,611 B |        14,010 B |
| DefaultsAsync | String[0] |   38.594 us |  29.371 us |         25,611 B |        14,010 B |
|   MinimalSync | String[0] |    6.905 us |   3.366 us |         12,201 B |         6,753 B |
|  MinimalAsync | String[0] |    7.084 us |   3.453 us |         12,201 B |         6,753 B |
|  DefaultsSync | String[4] |   44.543 us |  33.569 us |         30,016 B |        17,432 B |
| DefaultsAsync | String[4] |   45.152 us |  33.727 us |         30,016 B |        17,432 B |
|   MinimalSync | String[4] |   10.513 us |   5.739 us |         16,608 B |        10,176 B |
|  MinimalAsync | String[4] |   10.398 us |   5.888 us |         16,608 B |        10,176 B |