@jonsequitur 2021-12-10
This is a great improvement! Some of these changes were already on my to-do list so thank you for doing this.

One change that I'm wondering about is the addition of the `IMatchable` interface to support name-based matching. I'd like to investigate an alternate approach.

We've recently moved the name-based binding out into a separate package (System.CommandLine.NamingConventionBinder). The name matching functionality is partly there to support that but also to support validation of the parser, e.g. detect naming collisions that would make a command line unparseable. 

With that in mind, can you think of an approach that avoids introducing an interface?

On a possibly related note, we've also discussed [removing `ICommand`, `IArgument`, and `IOption`](https://github.com/dotnet/command-line-api/issues/1402). They were introduced when we made the parser mutable in order to avoid writing code to gently discourage mutating symbols once the parser is constructed. I don't believe they add value, though, and we don't expect people to implement them.