@adamsitnik 2021-12-13
> can you think of an approach that avoids introducing an interface?

Sure! I'll see what I can do and get back to you.

> On a possibly related note, we've also discussed removing ICommand, IArgument, and IOption

I am in favor of removing every abstraction that is implemented by a single type and|or is never implemented by the users. 👍